Adding functions to hammburger menu and sidebar

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -98,6 +99,10 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+        if($subCategory>0){
+            return response(['status'=> 'error', 'message' => 'Item contains sub-item, to delete this item delelte its sub-items first']);
+        }
         $category->delete();
 
         return response(['status' => 'success', 'Deleted Successfully!']);
